Fix incorrect pair type comparison (#238)
For some of my project calls the interpreted failed with `set is not sorted` error for `set` containing pairs of type `address * nat`. This fix helped me and it seems logical: before it assumed if earlier element is smaller it immediately considers whole pair as smaller. Where as now it makes sure every item is actually less than or equal to the other one to be sure whole pair is smaller.